feat(agent,tests,deps): decompose large modules, improve test coverage, remove deprecated dep
- Extract tool_call_parser.rs (~1600 lines) and tool_filter.rs (~190 lines)
  from agent/loop_.rs, reducing it from 6989 to ~5200 lines
- Replace deprecated serde_yaml dependency with lightweight manual YAML
  frontmatter parser in skills/mod.rs (~70 lines)
- Add ~55 new unit tests across 5 previously undertested modules:
  - multimodal.rs: MIME detection, validation, image markers, data URIs
  - commands/self_test.rs: check constructors, config/workspace/sqlite checks
  - commands/update.rs: version comparison, asset URL matching
  - peripherals/serial.rs: serial port path allowlisting, GPIO tool schema
  - hardware/registry.rs: board lookups, VID/PID matching, duplicate detection
- Apply cargo fmt formatting fixes
