+++ Cargo.toml
@@ -54,7 +54,7 @@ matrix-sdk = { version = "0.16", optional = true, default-features = false, feat
 serde = { version = "1.0", default-features = false, features = ["derive"] }
 serde_json = { version = "1.0", default-features = false, features = ["std"] }
 serde_ignored = "0.1"
-serde_yaml = "0.9"
+
 
 # Config
 directories = "6.0"
