+++ src/commands/update.rs
@@ -273,4 +273,88 @@ mod tests {
         assert!(!version_is_newer("0.4.3", "0.4.3"));
         assert!(version_is_newer("1.0.0", "2.0.0"));
     }
+
+    #[test]
+    fn version_is_newer_handles_major_version_bump() {
+        assert!(version_is_newer("0.9.9", "1.0.0"));
+        assert!(!version_is_newer("1.0.0", "0.9.9"));
+    }
+
+    #[test]
+    fn version_is_newer_handles_patch_only() {
+        assert!(version_is_newer("1.2.3", "1.2.4"));
+        assert!(!version_is_newer("1.2.4", "1.2.3"));
+    }
+
+    #[test]
+    fn version_is_newer_handles_non_numeric_components() {
+        // Non-numeric parts are filtered out by parse().ok()
+        assert!(version_is_newer("0.4.3", "0.5.0-rc1"));
+    }
+
+    #[test]
+    fn version_is_newer_handles_different_length() {
+        // "1.0" vs "1.0.1" — shorter version compares element-wise
+        assert!(version_is_newer("1.0", "1.0.1"));
+    }
+
+    #[test]
+    fn find_asset_url_returns_none_for_empty_assets() {
+        let release = serde_json::json!({ "assets": [] });
+        assert!(find_asset_url(&release).is_none());
+    }
+
+    #[test]
+    fn find_asset_url_returns_none_for_missing_assets() {
+        let release = serde_json::json!({ "tag_name": "v0.5.0" });
+        assert!(find_asset_url(&release).is_none());
+    }
+
+    #[test]
+    fn find_asset_url_finds_matching_asset() {
+        let target = if cfg!(target_os = "macos") {
+            if cfg!(target_arch = "aarch64") {
+                "aarch64-apple-darwin"
+            } else {
+                "x86_64-apple-darwin"
+            }
+        } else if cfg!(target_os = "linux") {
+            if cfg!(target_arch = "aarch64") {
+                "aarch64-unknown-linux"
+            } else {
+                "x86_64-unknown-linux"
+            }
+        } else {
+            return; // skip on unsupported platforms
+        };
+
+        let release = serde_json::json!({
+            "assets": [
+                {
+                    "name": format!("zeroclaw-{target}"),
+                    "browser_download_url": format!("https://github.com/example/releases/download/v0.5.0/zeroclaw-{target}")
+                }
+            ]
+        });
+
+        let url = find_asset_url(&release);
+        assert!(url.is_some());
+        assert!(url.unwrap().contains(target));
+    }
+
+    #[test]
+    fn find_asset_url_skips_non_matching_assets() {
+        let release = serde_json::json!({
+            "assets": [
+                {
+                    "name": "zeroclaw-some-other-platform",
+                    "browser_download_url": "https://example.com/download"
+                }
+            ]
+        });
+
+        // This may or may not match depending on the test platform,
+        // but we test that it doesn't panic
+        let _ = find_asset_url(&release);
+    }
 }
