+++ src/multimodal.rs
@@ -590,4 +590,201 @@ mod tests {
         );
         assert_eq!(refs.len(), 1);
     }
+
+    // ── MIME detection tests ──────────────────────────────────────────
+
+    #[test]
+    fn mime_from_magic_detects_png() {
+        let png_bytes = [0x89, b'P', b'N', b'G', b'\r', b'\n', 0x1a, b'\n'];
+        assert_eq!(mime_from_magic(&png_bytes), Some("image/png"));
+    }
+
+    #[test]
+    fn mime_from_magic_detects_jpeg() {
+        let jpeg_bytes = [0xff, 0xd8, 0xff, 0xe0, 0x00, 0x10];
+        assert_eq!(mime_from_magic(&jpeg_bytes), Some("image/jpeg"));
+    }
+
+    #[test]
+    fn mime_from_magic_detects_gif87a() {
+        assert_eq!(mime_from_magic(b"GIF87a\x00\x00"), Some("image/gif"));
+    }
+
+    #[test]
+    fn mime_from_magic_detects_gif89a() {
+        assert_eq!(mime_from_magic(b"GIF89a\x00\x00"), Some("image/gif"));
+    }
+
+    #[test]
+    fn mime_from_magic_detects_webp() {
+        let mut webp_bytes = b"RIFF\x00\x00\x00\x00WEBP".to_vec();
+        webp_bytes.extend_from_slice(&[0u8; 20]);
+        assert_eq!(mime_from_magic(&webp_bytes), Some("image/webp"));
+    }
+
+    #[test]
+    fn mime_from_magic_detects_bmp() {
+        assert_eq!(mime_from_magic(b"BM\x00\x00\x00\x00"), Some("image/bmp"));
+    }
+
+    #[test]
+    fn mime_from_magic_returns_none_for_unknown() {
+        assert_eq!(mime_from_magic(b"\x00\x00\x00\x00"), None);
+        assert_eq!(mime_from_magic(b""), None);
+        assert_eq!(mime_from_magic(b"PK"), None);
+    }
+
+    #[test]
+    fn mime_from_extension_covers_all_image_types() {
+        assert_eq!(mime_from_extension("png"), Some("image/png"));
+        assert_eq!(mime_from_extension("jpg"), Some("image/jpeg"));
+        assert_eq!(mime_from_extension("jpeg"), Some("image/jpeg"));
+        assert_eq!(mime_from_extension("webp"), Some("image/webp"));
+        assert_eq!(mime_from_extension("gif"), Some("image/gif"));
+        assert_eq!(mime_from_extension("bmp"), Some("image/bmp"));
+        assert_eq!(mime_from_extension("svg"), None);
+        assert_eq!(mime_from_extension("txt"), None);
+    }
+
+    #[test]
+    fn mime_from_extension_is_case_insensitive() {
+        assert_eq!(mime_from_extension("PNG"), Some("image/png"));
+        assert_eq!(mime_from_extension("Jpg"), Some("image/jpeg"));
+    }
+
+    #[test]
+    fn normalize_content_type_extracts_mime() {
+        assert_eq!(
+            normalize_content_type("image/png"),
+            Some("image/png".to_string())
+        );
+        assert_eq!(
+            normalize_content_type("image/jpeg; charset=utf-8"),
+            Some("image/jpeg".to_string())
+        );
+        assert_eq!(
+            normalize_content_type("IMAGE/PNG"),
+            Some("image/png".to_string())
+        );
+        assert_eq!(normalize_content_type(""), None);
+    }
+
+    #[test]
+    fn detect_mime_prefers_header_over_extension_and_magic() {
+        let png_bytes = [0x89, b'P', b'N', b'G', b'\r', b'\n', 0x1a, b'\n'];
+        // Header says jpeg, magic says png → header wins
+        let result = detect_mime(None, &png_bytes, Some("image/jpeg"));
+        assert_eq!(result, Some("image/jpeg".to_string()));
+    }
+
+    #[test]
+    fn detect_mime_falls_back_to_extension() {
+        let result = detect_mime(Some(Path::new("/tmp/test.webp")), b"\x00\x00\x00\x00", None);
+        assert_eq!(result, Some("image/webp".to_string()));
+    }
+
+    #[test]
+    fn detect_mime_falls_back_to_magic_bytes() {
+        let jpeg_bytes = [0xff, 0xd8, 0xff];
+        let result = detect_mime(None, &jpeg_bytes, None);
+        assert_eq!(result, Some("image/jpeg".to_string()));
+    }
+
+    #[test]
+    fn validate_mime_accepts_allowed_types() {
+        assert!(validate_mime("test", "image/png").is_ok());
+        assert!(validate_mime("test", "image/jpeg").is_ok());
+        assert!(validate_mime("test", "image/gif").is_ok());
+        assert!(validate_mime("test", "image/webp").is_ok());
+        assert!(validate_mime("test", "image/bmp").is_ok());
+    }
+
+    #[test]
+    fn validate_mime_rejects_disallowed_types() {
+        assert!(validate_mime("test", "image/svg+xml").is_err());
+        assert!(validate_mime("test", "application/pdf").is_err());
+        assert!(validate_mime("test", "text/plain").is_err());
+    }
+
+    #[test]
+    fn validate_size_enforces_limit() {
+        assert!(validate_size("test", 100, 200).is_ok());
+        assert!(validate_size("test", 200, 200).is_ok());
+        assert!(validate_size("test", 201, 200).is_err());
+    }
+
+    // ── Data URI handling tests ──────────────────────────────────────
+
+    #[test]
+    fn normalize_data_uri_accepts_valid_base64() {
+        let uri = "data:image/png;base64,iVBORw0KGgo=";
+        let result = normalize_data_uri(uri, 1024 * 1024);
+        assert!(result.is_ok());
+        assert!(result.unwrap().starts_with("data:image/png;base64,"));
+    }
+
+    #[test]
+    fn normalize_data_uri_rejects_non_base64() {
+        let uri = "data:image/png,rawdata";
+        let result = normalize_data_uri(uri, 1024 * 1024);
+        assert!(result.is_err());
+    }
+
+    #[test]
+    fn normalize_data_uri_rejects_unsupported_mime() {
+        let uri = "data:text/plain;base64,aGVsbG8=";
+        let result = normalize_data_uri(uri, 1024 * 1024);
+        assert!(result.is_err());
+    }
+
+    #[test]
+    fn count_image_markers_counts_correctly() {
+        let msgs = vec![
+            ChatMessage::user("[IMAGE:/a.png] text [IMAGE:/b.jpg]".to_string()),
+            ChatMessage::user("no images here".to_string()),
+            ChatMessage::user("[IMAGE:/c.gif]".to_string()),
+        ];
+        assert_eq!(count_image_markers(&msgs), 3);
+    }
+
+    #[test]
+    fn contains_image_markers_detects_presence() {
+        let with = vec![ChatMessage::user("[IMAGE:/a.png] text".to_string())];
+        let without = vec![ChatMessage::user("no images here".to_string())];
+        assert!(contains_image_markers(&with));
+        assert!(!contains_image_markers(&without));
+    }
+
+    #[test]
+    fn extract_ollama_image_payload_passes_through_non_data_uri() {
+        // Non-data URIs are treated as raw base64 payloads and returned as-is.
+        assert_eq!(
+            extract_ollama_image_payload("https://example.com/img.png"),
+            Some("https://example.com/img.png".to_string())
+        );
+        assert_eq!(
+            extract_ollama_image_payload("/tmp/local.png"),
+            Some("/tmp/local.png".to_string())
+        );
+        // Empty strings are rejected.
+        assert!(extract_ollama_image_payload("").is_none());
+        assert!(extract_ollama_image_payload("   ").is_none());
+    }
+
+    #[tokio::test]
+    async fn normalize_local_image_rejects_nonexistent_file() {
+        let result = normalize_local_image("/nonexistent/path/image.png", 1024 * 1024).await;
+        assert!(result.is_err());
+    }
+
+    #[tokio::test]
+    async fn prepare_messages_passes_through_no_image_messages() {
+        let messages = vec![ChatMessage::user("just text, no images".to_string())];
+        let prepared = prepare_messages_for_provider(&messages, &MultimodalConfig::default())
+            .await
+            .unwrap();
+        assert!(!prepared.contains_images);
+        assert_eq!(prepared.messages.len(), 1);
+        assert_eq!(prepared.messages[0].content, "just text, no images");
+    }
 }
